move topappbar to views
